Create minimum viable product with console output
This commit is a very stripped back version of
the final project. Tasks can be displayed via the
console, they can be displayed based on whether
they contained a particular property, in this case
'Home' has been hard-coded to sort the project
property. Tasks can be added, removed and updated
based on their ID - for which any specified
property can be updated, this will be useful later
for with updating details of the tasks and their
'completed' state checkbox.